broke out functions for epc, new species
